+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Deezer Changelog
 ================
 
+  * Fix track title and artist in queue view. Issue: tiliado/nuvolaruntime#609
+
 3.2 - February 24th, 2019
 -------------------------